remove malloc from solo5
